Add additional soil sensor impl and rework pmindp-esp32-thread to support generic sensor type
