+++ pmindp-esp32-thread/Cargo.toml
@@ -5,6 +5,8 @@ edition = "2021"
 authors = ["nand-nor <a13xandra.cliff0rd@gmail.com>"]
 
 [dependencies]
+cfg-if = "1.0.0"
+embedded-hal = {version = "1.0.0"} 
 esp-hal = {version= "0.19.0", git = "https://github.com/esp-rs/esp-hal.git"}
 esp-backtrace = { version="0.13.0", git = "https://github.com/esp-rs/esp-hal.git", features = ["panic-handler", "exception-handler", "println"]}
 esp-println       = { version= "0.10.0", git = "https://github.com/esp-rs/esp-hal.git", features = ["log"]}
@@ -17,10 +19,13 @@ no-std-net = {version= "0.6.0"}
 critical-section = {version= "1.1.0"}
 smart-leds = {version= "0.4.0" }
 esp-openthread = { git = "https://github.com/nand-nor/esp-openthread.git", rev="d18b689"}
-pmindp-sensor = {  path="../pmindp-sensor"}
 coap-lite = {version="0.12.0", features=["udp"],default-features=false}
+pmindp-sensor = {  path="../pmindp-sensor"}
 
 [features]
-default = ["esp32c6"]
+default = ["esp32c6","atsamd10"]
 esp32c6 = [ "esp-hal/esp32c6", "esp-ieee802154/esp32c6", "esp-openthread/esp32c6", "esp-backtrace/esp32c6", "esp-println/esp32c6", "esp-hal-smartled/esp32c6" ]
 esp32h2 = [ "esp-hal/esp32h2", "esp-ieee802154/esp32h2", "esp-openthread/esp32h2", "esp-backtrace/esp32h2", "esp-println/esp32h2", "esp-hal-smartled/esp32h2" ]
+probe-circuit = []
+atsamd10 = []
+st0160 = []
