+++ pmindp-esp32-thread/src/bin/main.rs
@@ -9,13 +9,20 @@ use esp_hal::{
     peripherals::Peripherals,
     prelude::*,
     system::SystemControl,
-    timer::{
-        systimer::SystemTimer,
-        timg::TimerGroup,
-    },
+    timer::{systimer::SystemTimer, timg::TimerGroup},
 };
+
+#[cfg(feature = "atsamd10")]
+use esp_hal::i2c::I2C;
+
+#[cfg(feature = "probe-circuit")]
+use esp_hal::{
+    delay::Delay,
+    gpio::{Level, Output},
+};
+
 use esp_println::println;
-use pmindp_esp32_thread::{init_heap, Esp32Platform};
+use pmindp_esp32_thread::init_heap;
 
 use esp_ieee802154::Ieee802154;
 
@@ -32,21 +39,62 @@ fn main() -> ! {
     let mut ieee802154 = Ieee802154::new(peripherals.IEEE802154, &mut peripherals.RADIO_CLK);
     let io = Io::new(peripherals.GPIO, peripherals.IO_MUX);
 
-    let mut platform = Esp32Platform::new(
-        &mut ieee802154,
-        &clocks,
-        systimer.alarm0,
-        peripherals.I2C0,
-        TimerGroup::new(
-            peripherals.TIMG0,
-            &clocks,
-        ),
-        peripherals.RMT,
-        io.pins.gpio8,
-        io.pins.gpio5,
-        io.pins.gpio6,
-        peripherals.RNG,
-    );
+    cfg_if::cfg_if! {
+        if #[cfg(feature="atsamd10")] {
+
+            let sensor = pmindp_esp32_thread::ATSAMD10 {
+                i2c: I2C::new(
+                    peripherals.I2C0,
+                    io.pins.gpio5,
+                    io.pins.gpio6,
+                    400.kHz(),
+                    &clocks,
+                ),
+                temp_delay: 2000,
+                moisture_delay: 5000,
+                address: 0x36,
+            };
+
+            let mut platform = pmindp_esp32_thread::init(
+                &mut ieee802154,
+                &clocks,
+                systimer.alarm0,
+                TimerGroup::new(peripherals.TIMG0, &clocks),
+                peripherals.RMT,
+                io.pins.gpio8,
+                peripherals.RNG,
+                sensor
+            );
+        } else if #[cfg(feature="probe-circuit")] {
+
+            let sensor = pmindp_esp32_thread::ProbeCircuit::new(
+                Output::new(
+                    io.pins.gpio6,
+                    Level::Low
+                ),
+                io.pins.gpio2,
+                peripherals.ADC1,
+                Delay::new(&clocks)
+            );
+
+            let mut platform = pmindp_esp32_thread::init(
+                &mut ieee802154,
+                &clocks,
+                systimer.alarm0,
+                TimerGroup::new(peripherals.TIMG0, &clocks),
+                peripherals.RMT,
+                io.pins.gpio8,
+                peripherals.RNG,
+                sensor
+            );
+
+        } else {
+            log::error!("No sensor target specified!");
+            loop {
+
+            }
+        }
+    }
 
     loop {
         // this will enter a loop where if it ever breaks,
