+++ pmindp-esp32-thread/src/sensor/mod.rs
@@ -1,0 +1,12 @@
+mod atsamd10;
+mod probe_circuit;
+mod st0160;
+
+//#[cfg(all(feature = "probe-circuit", not(feature = "atsamd10", feature = "st0160")))]
+pub use probe_circuit::ProbeCircuit;
+
+//#[cfg(all(feature = "atsamd10", not(feature = "probe-circuit", feature = "st0160")))]
+pub use atsamd10::ATSAMD10;
+
+//#[cfg(all(feature = "st0160", not(feature = "atsamd10", feature = "probe-circuit")))]
+pub use st0160::ST0160;
