+++ pmindp-esp32-thread/src/sensor/st0160.rs
@@ -1,0 +1,4 @@
+/// Sunfounder capacitive soil moisture sensor
+/// https://www.digikey.com/en/products/detail/sunfounder/ST0160/22116813
+
+pub struct ST0160 {}
